# reduce chunk size of Hessian in Landau example to improve latency and performance