12 is a very big chunk size for a hessian because you end up with Duals of 12^2 + 12 + 1 numbers in them.

Running with the following model

```julia
model = LandauModel(α, G, (10, 10, 2), left, right, element_potential);
```

this changes the times from

```
First run
 36.561361 seconds (16.29 M allocations: 1.118 GiB, 1.40% gc time, 90.71% compilation time)
Second run
  2.889448 seconds (63.62 k allocations: 321.635 MiB, 1.80% gc time)
```

to
```
First run:
  4.490308 seconds (6.43 M allocations: 494.372 MiB, 1.63% gc time, 63.58% compilation time)
Second run
  1.673789 seconds (90.01 k allocations: 118.406 MiB, 0.73% gc time)
```

so both a latency and a perf improvement.