@termi-official 2024-08-01
The change is fine, but why do we use AD on the full element instead of AD for each quadrature point here in first place?