@KristofferC 2024-08-01
I guess because we use AD to compute the derivative of the actual residual and not just the derivative of a "constitutive driver". I'm not sure it will matters much since all the quadrature loop does is accumulate the contributions.